view: Cast playlist position to gint before using it to set the property
celluloid_view_set_playlist_pos() accepts the position as a gint64 but
the playlist-pos property is an int.

Fixes #660.

+++ src/celluloid-view.c
@@ -1759,7 +1759,7 @@ celluloid_view_update_playlist(CelluloidView *view, GPtrArray *playlist)
 void
 celluloid_view_set_playlist_pos(CelluloidView *view, gint64 pos)
 {
-	g_object_set(view, "playlist-pos", pos, NULL);
+	g_object_set(view, "playlist-pos", (gint)pos, NULL);
 }
 
 void
